reorg to allow invoking bad kmer finder as static method
